basic_app_using_lchain/requirements.txt
langchain
ipykernel
python-dotenv
langchain_community 
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain_ollama
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette